requirements.txt
streamlit==1.37.0
openai>=1.40.0
tiktoken
chromadb==0.5.5
pydantic>=2.7
pyvis==0.3.2
networkx>=3.3
python-dotenv>=1.0.1
tenacity>=8.3.0
